+++ psi4/src/psi4/dfmp2/corr_grad.h
@@ -171,15 +171,15 @@ protected:
     void build_AB_x_terms();
     void build_Amn_x_terms();
 
-    void fitting_helper(SharedMatrix J, unsigned int file, const std::string& label, size_t naux, size_t nij, size_t memory);
-    void UV_helper(SharedMatrix V, double c, unsigned int file, const std::string& label, size_t naux, size_t nij, size_t memory);
+    void fitting_helper(SharedMatrix J, size_t file, const std::string& label, size_t naux, size_t nij, size_t memory);
+    void UV_helper(SharedMatrix V, double c, size_t file, const std::string& label, size_t naux, size_t nij, size_t memory);
 
     /// File number for Alpha (Q|mn) tensor
-    unsigned int unit_a_;
+    size_t unit_a_;
     /// File number for Beta (Q|mn) tensor
-    unsigned int unit_b_;
+    size_t unit_b_;
     /// File number for J tensors
-    unsigned int unit_c_;
+    size_t unit_c_;
 
 public:
     DFCorrGrad(std::shared_ptr<BasisSet> primary, std::shared_ptr<BasisSet> auxiliary);
@@ -201,17 +201,17 @@ public:
      * Which file number should the Alpha (Q|mn) integrals go in
      * @param unit Unit number
      */
-    void set_unit_a(unsigned int unit) { unit_a_ = unit; }
+    void set_unit_a(size_t unit) { unit_a_ = unit; }
     /**
      * Which file number should the Beta (Q|mn) integrals go in
      * @param unit Unit number
      */
-    void set_unit_b(unsigned int unit) { unit_b_ = unit; }
+    void set_unit_b(size_t unit) { unit_b_ = unit; }
     /**
      * Which file number should the J tensors go in
      * @param unit Unit number
      */
-    void set_unit_c(unsigned int unit) { unit_c_ = unit; }
+    void set_unit_c(size_t unit) { unit_c_ = unit; }
 
     /**
      * What number of threads to compute integrals on
