+++ psi4/src/psi4/libdpd/file4_init.cc
@@ -58,7 +58,7 @@ int DPD::file4_init(dpdfile4 *File, int filenum, int irrep, int pqnum,
 {
     int i;
     int maxrows, rowtot, coltot;
-    unsigned int priority;
+    size_t priority;
     dpd_file4_cache_entry *this_entry;
     psio_address irrep_ptr;
